updated db helper such that the comment will not be overwritten, but instead be an aggregation of comments to see the history
updated edit task so that comment will not be pulled to the edit task field, and the field will show blank

updated edit task so that when a comment is saved, the date and who ever is logged in that left the comment will be saved as well. in an aggregation manner

update add task with the same logic as edit task

created a new activity to display the comment history. shows the date of the comment and who put the comment